% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package-class.R
\docType{class}
\name{BubleyDyerGenerator-class}
\alias{BubleyDyerGenerator-class}
\title{An S4 class to represent the linear extension generator based on the Bubley-Dyer procedure.}
\description{
An S4 class to represent the linear extension generator based on the Bubley-Dyer procedure.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{an external pointer to C++ data}

\item{\code{seed}}{the 64-bit seed actually used, as a string of decimal digits}
}}

