% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package-class.R
\docType{class}
\name{BubleyDyerSeparationGenerator-class}
\alias{BubleyDyerSeparationGenerator-class}
\title{An S4 class to represent function separation based on the Bubley-Dyer procedure.}
\description{
An S4 class to represent function separation based on the Bubley-Dyer procedure.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{an external pointer to a C++ data}

\item{\code{types}}{list of separation to be computed}

\item{\code{seed}}{the 64-bit seed actually used, as a string of decimal digits}
}}

