% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dominates.R
\name{Dominates}
\alias{Dominates}
\title{Checking whether one element dominates another.}
\usage{
Dominates(poset, element1, element2)
}
\arguments{
\item{poset}{An object of S4 class \code{POSet}.
Argument \code{poset} must be created by using any function contained in the package aimed at building object of S4 class \code{POSet}
(e.g. \code{\link[=POSet]{POSet()}}, \code{\link[=LinearPOSet]{LinearPOSet()}}, \code{\link[=ProductPOSet]{ProductPOSet()}}, ...) .}

\item{element1}{A character string or vector (the name(s) of poset element(s)).}

\item{element2}{A character string or vector (the name(s) of poset element(s)).}
}
\value{
A boolean value.
}
\description{
Given two elements \eqn{a} and \eqn{b} of \eqn{V}, checks whether \eqn{a\leq b} in poset \eqn{(V,\leq)}.
}
\examples{
elems <- c("a", "b", "c", "d")

dom <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)

pos <- POSet(elements = elems, dom = dom)

chk <- Dominates(pos, "a", "d")

}
