% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FencePOSet.R
\name{FencePOSet}
\alias{FencePOSet}
\title{Building fences}
\usage{
FencePOSet(elements, orientation = "upFirst")
}
\arguments{
\item{elements}{A character vector (the names of the fence elements).}

\item{orientation}{Either "upFirst" (the first element dominates the second) or
"downFirst" (the second element dominates the first). Default is "upFirst".}
}
\value{
A fence, an object of S4 class \code{POSet}.
}
\description{
Builds a fence from an unordered collection of elements.
}
\examples{
elems <- c("a", "b", "c", "d", "e")
fence <- FencePOSet(elems, orientation = "upFirst")

}
