% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package-class.R
\docType{class}
\name{FromPOSet-class}
\alias{FromPOSet-class}
\title{An S4 class to represent a virtual class for POSet extention.}
\description{
An S4 class to represent a virtual class for POSet extention.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{an external pointer to C++ data}
}}

