% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInBetweenness.R
\name{FuzzyInBetweenness}
\alias{FuzzyInBetweenness}
\title{Fuzzy in-betweenness array computation}
\usage{
FuzzyInBetweenness(
  dom,
  norm,
  conorm,
  types = c("symmetric", "asymmetricLower", "asymmetricUpper")
)
}
\arguments{
\item{dom}{A square numeric matrix representing the dominance degree between pairs of poset elements.
Row and column names of \code{dom} are interpreted as the labels of the poset elements.
\code{dom} can be computed by using functions such as \code{\link[=BLSDominance]{BLSDominance()}}, \code{\link[=BubleyDyerMRP]{BubleyDyerMRP()}}, or \code{\link[=ExactMRP]{ExactMRP()}}.}

\item{norm}{An R function defining the t-norm.}

\item{conorm}{An R function defining the t-conorm.}

\item{types}{Character vector specifying the types of in-betweenness to be computed.
Possible choices are: \code{"symmetric"}, \code{"asymmetricLower"}, and \code{"asymmetricUpper"}.
Multiple choices can be provided (e.g., \code{types = c("symmetric", "asymmetricLower")}).
Defaults to \code{"symmetric"}.
For details on the definition of symmetric and asymmetric in-betweenness, see Fattore et al. (2024).}
}
\value{
A list of three-dimensional arrays, one for each type of in-betweenness selected by \code{types}.
The array element at position \eqn{[i,j,k]} represents \eqn{finb_{p_i,p_j,p_k}} for symmetric in-betweenness,
\eqn{finb_{p_i<p_j<p_k}} for asymmetricLower in-betweenness, and \eqn{finb_{p_k<p_j<p_i}} for asymmetricUpper in-betweenness.
}
\description{
Starting from a poset dominance matrix, computes in-betweenness arrays using a user-supplied t-norm and t-conorm.
}
\examples{
el <- c("a", "b", "c", "d")

dom_list <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)

pos <- POSet(elements = el, dom = dom_list)

BLS <- BLSDominance(pos)

tnorm   <- function(x, y) { x * y }

tconorm <- function(x, y) { x + y - x * y }

FinB <- FuzzyInBetweenness(BLS, norm = tnorm, conorm = tconorm,
                            types = c("symmetric", "asymmetricLower"))

}
\references{
Fattore, M., De Capitani, L., Avellone, A., and Suardi, A. (2024).
A fuzzy posetic toolbox for multi-criteria evaluation on ordinal data systems.
Annals of Operations Research, https://doi.org/10.1007/s10479-024-06352-3.
}
