% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzySeparationProbabilistic.R
\name{FuzzySeparationProbabilistic}
\alias{FuzzySeparationProbabilistic}
\title{Fuzzy separation matrix computation with Product t-norm and Probabilistic-sum t-conorm}
\usage{
FuzzySeparationProbabilistic(
  dom,
  types = c("symmetric", "asymmetricLower", "asymmetricUpper", "vertical", "horizontal")
)
}
\arguments{
\item{dom}{A square numeric matrix representing the dominance degrees between pairs of poset elements.
Row and column names of \code{dom} are interpreted as the labels of the poset elements.
\code{dom} can be computed by using functions such as \code{\link[=BLSDominance]{BLSDominance()}}, \code{\link[=BubleyDyerMRP]{BubleyDyerMRP()}}, or \code{\link[=ExactMRP]{ExactMRP()}}.}

\item{types}{Character vector specifying the types of fuzzy separation to be computed.
Possible choices are: \code{"symmetric"}, \code{"asymmetricLower"}, \code{"asymmetricUpper"},
\code{"vertical"}, and \code{"horizontal"}. Multiple choices can be provided
(e.g., \code{types = c("symmetric", "vertical")}). Defaults to \code{"symmetric"}.
For details on the definition of symmetric, asymmetric, vertical, and horizontal separations, see Fattore et al. (2024).}
}
\value{
A list of the required fuzzy separation matrices.
}
\description{
Starting from a poset dominance matrix, computes fuzzy separation matrices using Product t-norm and Probabilistic-sum t-conorm.
}
\examples{
el <- c("a", "b", "c", "d")

dom_list <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)

pos <- POSet(elements = el, dom = dom_list)

BLS <- BLSDominance(pos)

FSep <- FuzzySeparationProbabilistic(BLS, types = c("symmetric", "vertical"))

}
\references{
Fattore, M., De Capitani, L., Avellone, A., and Suardi, A. (2024).
A fuzzy posetic toolbox for multi-criteria evaluation on ordinal data systems.
Annals of Operations Research, https://doi.org/10.1007/s10479-024-06352-3.
}
