% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncomparabilitySetOf.R
\name{IncomparabilitySetOf}
\alias{IncomparabilitySetOf}
\title{Extracting the incomparability set of a poset element}
\usage{
IncomparabilitySetOf(poset, element)
}
\arguments{
\item{poset}{An object of S4 class \code{POSet}.
Argument \code{poset} must be created by using any function contained in the package aimed at building object of S4 class \code{POSet}
(e.g. \code{\link[=POSet]{POSet()}}, \code{\link[=LinearPOSet]{LinearPOSet()}}, \code{\link[=ProductPOSet]{ProductPOSet()}}, ...).}

\item{element}{A character string (the name of a single poset element).}
}
\value{
A character vector containing the names of the poset elements incomparable with the input element.
}
\description{
Extracts the elements incomparable with the input element in the poset.
}
\examples{
elems <- c("a", "b", "c", "d")

dom <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)

pos <- POSet(elements = elems, dom = dom)
incmp <- IncomparabilitySetOf(pos, "a")

}
