% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsAntisymmetric.R
\name{IsAntisymmetric}
\alias{IsAntisymmetric}
\title{Checking binary relation antisymmetry.}
\usage{
IsAntisymmetric(rel)
}
\arguments{
\item{rel}{A two-column character matrix, each row comprising an element (pair) of the binary relation.}
}
\value{
A logical value: TRUE if the relation is antisymmetric, FALSE otherwise.
}
\description{
Checks whether the input binary relation is antisymmetric.
}
\examples{
rel <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d",
  "a", "a"
), ncol = 2, byrow = TRUE)

chk <- IsAntisymmetric(rel)

}
