% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsComparableWith.R
\name{IsComparableWith}
\alias{IsComparableWith}
\title{Checking comparability between two elements of a poset.}
\usage{
IsComparableWith(poset, element1, element2)
}
\arguments{
\item{poset}{An object of S4 class \code{POSet}.
Argument \code{poset} must be created by using any function contained in the package aimed at building object of S4 class \code{POSet}
(e.g. \code{\link[=POSet]{POSet()}}, \code{\link[=LinearPOSet]{LinearPOSet()}}, \code{\link[=ProductPOSet]{ProductPOSet()}}, ...) .}

\item{element1}{A character vector (the names of poset elements).}

\item{element2}{A character vector (the names of poset elements).}
}
\value{
A logical value (or a logical vector if inputs are vectors).
}
\description{
Checks whether two elements \eqn{a} and \eqn{b} of \eqn{V} are comparable in the input poset \eqn{(V,\leq)}.
}
\examples{
elems <- c("a", "b", "c", "d")

dom <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)

pos <- POSet(elements = elems, dom = dom)

chk <- IsComparableWith(pos, "a", "d")

}
