% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsSymmetric.R
\name{IsSymmetric}
\alias{IsSymmetric}
\title{Checking binary relation symmetry.}
\usage{
IsSymmetric(rel)
}
\arguments{
\item{rel}{A two-columns character matrix, each row comprising an element (pair) of the binary relation.}
}
\value{
A boolean value.
}
\description{
Checks whether the input binary relation is symmetric.
}
\examples{
rel <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d",
  "b", "a",
  "b", "c",
  "d", "b"
), ncol = 2, byrow = TRUE)

chk<-IsSymmetric(rel)

}
