% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsTransitive.R
\name{IsTransitive}
\alias{IsTransitive}
\title{Checking binary relation transitivity.}
\usage{
IsTransitive(rel)
}
\arguments{
\item{rel}{A two-columns character matrix, each row comprising an element (pair) of the binary relation.}
}
\value{
A boolean value.
}
\description{
Checks whether the input relation is transitive.
}
\examples{
rel <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d",
  "a", "d",
  "c", "d"
), ncol = 2, byrow = TRUE)

chk<-IsTransitive(rel)

}
