% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsUpset.R
\name{IsUpset}
\alias{IsUpset}
\title{Checking upsets.}
\usage{
IsUpset(poset, elements)
}
\arguments{
\item{poset}{An object of S4 class \code{POSet}.
Argument \code{poset} must be created by using any function contained in the package aimed at building object of S4 class \code{POSet}
(e.g. \code{\link[=POSet]{POSet()}}, \code{\link[=LinearPOSet]{LinearPOSet()}}, \code{\link[=ProductPOSet]{ProductPOSet()}}, ...) .}

\item{elements}{A vector of character strings (the names of the input elements).}
}
\value{
A boolean value.
}
\description{
Checks whether the input elements form an upset, in the input poset.
}
\examples{
elems <- c("a", "b", "c", "d")

dom <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)

pos <- POSet(elements = elems, dom = dom)

chk <- IsUpset(pos, c("a", "b", "c"))


}
