% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package-class.R
\docType{class}
\name{LEGenerator-class}
\alias{LEGenerator-class}
\title{An S4 class to represent the exact linear extension generator.}
\description{
An S4 class to represent the exact linear extension generator.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{an external pointer to C++ data}
}}

