% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexSeparation.R
\name{LexSeparation}
\alias{LexSeparation}
\title{Separation matrices computation over the set of lexicographic linear extensions.}
\usage{
LexSeparation(
  nvar,
  deg,
  types = c("symmetric", "asymmetricLower", "asymmetricUpper", "vertical", "horizontal")
)
}
\arguments{
\item{nvar}{positive integer specifying the number \eqn{k} of ordinal variables.}

\item{deg}{parameter specifying the number of degrees of each variable. If all \eqn{k} variables have the same number \eqn{m} of degrees, it can be:
\enumerate{
\item the positive integer \eqn{m}. In this case variable degree labels are supposed to be the integers \eqn{1<2<...<m}
and columns and rows of computed separation matrices are named accordingly to this;
\item a character vector of length \eqn{m} specifying the variable degree labels (in this case columns and rows of computed separation matrices are
named accordingly to \code{deg}).
}

If the \eqn{k} variables have different number \eqn{(m_1,...,m_k)} of degrees, it can be:
\enumerate{
\item a length-\eqn{k} positive integers vector specifying the values of \eqn{m_1,...,m_k}. In this case variable degree labels for the \eqn{j}-th variable are
supposed to be the integers \eqn{1<2<...<m_j} and columns and rows of computed separation matrices are named accordingly to this;
\item a list of \eqn{k} character vectors. The \eqn{j}-th list element is a character vector of length \eqn{m_j} specifying the degree labels for the \eqn{j}-th variable
(in this case columns and rows of computed separation matrices are named accordingly to \code{deg}).
}}

\item{types}{Character vector specifying the types of separation matrix to be computed.
Possible choices are: \code{"symmetric"}, \code{"asymmetricLower"}, \code{"asymmetricUpper"},
\code{"vertical"}, and \code{"horizontal"}. Multiple choices can be provided
(e.g., \code{types = c("symmetric", "vertical")}). Defaults to \code{"symmetric"}.}
}
\value{
list of required separation matrices.
}
\description{
Considering the component-wise poset built starting from \eqn{k} ordinal variables,
computes separation matrices by analyzing all poset lexicographic linear extensions.
}
\examples{

#variables with common number of degrees
# default labels for variable degrees
nvar <- 3
deg  <- 4
LexSep <- LexSeparation(nvar=nvar, deg=deg, types = c("symmetric", "asymmetricLower"))

#user supplied variable degree labels
nvar <- 3
deg  <- c("a","b","c","d")
LexSep <- LexSeparation(nvar=nvar, deg=deg, types = c("symmetric", "asymmetricLower"))


#variables with different numbers of degrees
# default labels for variable degrees
nvar <- 3
deg  <- c(4,2,3)
LexSep <- LexSeparation(nvar=nvar, deg=deg, types = c("symmetric", "asymmetricLower"))

#user supplied variable degree labels
nvar <- 3
deg  <- list(c("a","b","c","d"),c("0","1"),c("x","y","z"))
LexSep <- LexSeparation(nvar=nvar, deg=deg, types = c("symmetric", "asymmetricLower"))

}
