% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package-class.R
\docType{class}
\name{LinearPOSet-class}
\alias{LinearPOSet-class}
\title{An S4 class to represent a Linear POSet.}
\description{
An S4 class to represent a Linear POSet.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{an external pointer to C++ data}
}}

