% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package-class.R
\docType{class}
\name{POSet-class}
\alias{POSet-class}
\title{An S4 class to represent a POSet.}
\description{
An S4 class to represent a POSet.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{an external pointer to C++ data}
}}

