% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSetFromDominance.R
\name{POSetFromDominance}
\alias{POSetFromDominance}
\title{Constructing a Partially Ordered Set from an Dominance Matrix.}
\usage{
POSetFromDominance(dominance_matrix)
}
\arguments{
\item{dominance_matrix}{A square numeric (0/1) or logical matrix representing
the partial order relation. An entry \verb{[i, j] == 1} (or \code{TRUE}) implies that
the element \code{i} is less than or equal to element \code{j} (\eqn{i \leq j}).}
}
\value{
An object of S4 class \code{POSet}.
}
\description{
Constructs an object of class \code{POSet} starting from an dominance matrix.
Element labels are extracted from the row/column names of the matrix.
If no names are provided, they are automatically generated.
}
\examples{
# Example with an unnamed logical matrix
inc_mat <- matrix(c(TRUE, FALSE, FALSE,
                    TRUE, TRUE,  FALSE,
                    TRUE, TRUE,  TRUE),
                  nrow = 3, byrow = TRUE)
pos <- POSetFromDominance(inc_mat)

# Example with a named numeric matrix
inc_mat_named <- matrix(c(1, 0, 1, 1), nrow = 2, byrow = TRUE)
rownames(inc_mat_named) <- colnames(inc_mat_named) <- c("A", "B")
pos_named <- POSetFromDominance(inc_mat_named)

}
