% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package-exports.R
\docType{package}
\name{poseticDataAnalysis-package}
\alias{poseticDataAnalysis}
\alias{poseticDataAnalysis-package}
\title{poseticDataAnalysis: Posetic Data Analysis}
\description{
Build and manipulate partially ordered sets (posets), to perform some data analysis on them and to implement multi-criteria decision making procedures. Several efficient ways for generating linear extensions are implemented, together with functions for building mutual ranking probabilities, incomparability, dominance and separation scores (Fattore, M., De Capitani, L., Avellone, A., Suardi, A. (2024). A fuzzy posetic toolbox for multi-criteria evaluation on ordinal data systems. ANNALS OF OPERATIONS RESEARCH \doi{10.1007/s10479-024-06352-3}).
}
\author{
\strong{Maintainer}: Alessandro Avellone \email{alessandro.avellone@unimib.it}

Authors:
\itemize{
  \item Alessandro Avellone \email{alessandro.avellone@unimib.it}
  \item Lucio De Capitani \email{lucio.decapitani1@unimib.it}
  \item Marco Fattore \email{marco.fattore@unimib.it}
}

}
